Share CountrySelect typeahead; use it in bulk Edit fields modal
The bulk Edit fields modal rendered a plain text input for the
country field, so the new "No country" sentinel was only one
click away in the single-row drawer — operators had to remember
to type `xx` to apply the sentinel across a selection.

Extracts the inline CountrySelect from PhotoDrawer to its own
Manage/CountrySelect.tsx (no behaviour change). PhotoDrawer
imports it as before; BulkActions's per-field render now branches
to CountrySelect for the country row, dropping the bare FieldInput
that didn't surface the sentinel.

Component API: optional `highlight` (PhotoDrawer's EXIF-override
border), optional `placeholder` (lets BulkActions reuse the
"<mixed>" placeholder shape when the selection's values differ).

The shared file makes future additions (alpha-3 fallback, etc.)
land in one place.
