+++ CHANGELOG.md
@@ -20,7 +20,7 @@
 - PhotoDrawer's read-only meta rows for `ID` and `Original filename` grow a small clipboard icon that copies the value via `navigator.clipboard.writeText`; flips to a check icon for 1.5s after a successful copy. Hidden when the clipboard API is unavailable (non-HTTPS dev origins). (closes #484)
 - PhotoDrawer country typeahead suggestions no longer get clipped by the photo's map view. The dropdown's `z-index: 10` lost to leaflet's internal panes (200–700) and to its control containers `.leaflet-top` / `-bottom` (1000) that wrap the +/− zoom buttons; bumped to 1100 so the typeahead floats above the map and its controls, while staying below the modal layer (2000) used by the bulk-action backdrops. (closes #488)
 - PhotoDrawer geocoded row no longer reads "Not geocoded yet" for photos taken in international waters. Display priority: shared `country / state / city` when city is set; else the first populated maritime field from Nominatim's address blob (`ocean` / `sea` / `bay` / `strait` / `gulf`); else "No location data" when `geocoded.noData` is true (Nominatim explicitly returned nothing); else the existing "Not geocoded yet" (daemon hasn't run). (closes #489)
-- PhotoDrawer country typeahead pins a "No country" option at the top of the dropdown (under the "Clear" option when a country is set). Matches the new `xx` sentinel from `country-sentinel.ts`. Saves the literal `xx` into `taken.location.country`. Gallery `MetadataPanel` and `Thumbnail` render the localised "No country" label instead of the raw `XX`, and skip the `FlagIcon` (no real flag exists for the sentinel). Stats country grouping and `SummaryModal` thread the `t` function through so the label renders correctly in their tooltips and tables. i18n key `country-sentinel-label` in en / fi / ja. (closes #486)
+- PhotoDrawer and the bulk Edit fields modal share a `CountrySelect` typeahead that pins a "No country" option at the top of the dropdown (under the "Clear" option when a country is set). Matches the new `xx` sentinel from `country-sentinel.ts`. Saves the literal `xx` into `taken.location.country`. Gallery `MetadataPanel` and `Thumbnail` render the localised "No country" label instead of the raw `XX`, and skip the `FlagIcon` (no real flag exists for the sentinel). Stats country grouping and `SummaryModal` thread the `t` function through so the label renders correctly in their tooltips and tables. i18n key `country-sentinel-label` in en / fi / ja. (closes #486)
 - Manage dashboard grows a "Catalogue audit" tile strip above the navigation tiles, one card per non-zero filter-chip predicate (orphan, duplicates, country mismatch, missing-{taken,coords,place,country,author,title,description}). Clicking a card lands on `/m/photos` with the matching chip pre-active. Cards with count zero are hidden so the strip reads as "what's still to fix". `state-code` is intentionally absent — the Photos page filter chip list excludes it (countries with no ISO-3166-2 subdivisions permanently match with no operator-editable fix), so a tile would land on an unfiltered view; CLI audit (`bin/photo.ts audit --missing state-code`) keeps the surface. Drawer save, delete, regeocode, and bulk actions invalidate the count query so the strip stays current as the operator works. (closes #454)
 - Manage Users list and edit form surface the new `user.name` column; Manage Groups + GroupEdit / GroupCreate rename "Title" → "Name"; access subject pickers append `— name` after the id when one is set. (part of #476)
 - Global Manage / Stats surfaces hide on hostname-bound instances. The server-side `requireUnscoped` guard (#386) rejects cross-gallery admin endpoints on a bound hostname with a 404, but the SPA was still rendering the Dashboard tiles, the UserMenu Statistics entry, and the landing-page Statistics shortcut card — clicks 404'd silently. New `useHostScope` hook reads the host-scope from the same primitive the breadcrumb / picker use; Manage Dashboard swaps its global tile set for the bound gallery's tiles (edit / photos / access) and single-gallery scope auto-redirects `/m` straight to `/m/g/<id>`; visiting any global Manage route bounces to `/m`; UserMenu drops the Statistics entry; landing-page drops the Statistics card; `/s` redirects to `/g`. (closes #462)
