+++ react-app/src/components/Manage/PhotoDrawer.tsx
@@ -18,11 +18,9 @@ import photosService, {
 } from "../../services/photos";
 import useKeyPress from "../../lib/keypress";
 import config from "../../lib/config";
-import {
-  COUNTRY_SENTINEL,
-  isCountrySentinel,
-} from "../../lib/country-sentinel";
+import { isCountrySentinel } from "../../lib/country-sentinel";
 import { useLangStore } from "../../stores";
+import CountrySelect from "./CountrySelect";
 import EditableMap from "./EditableMap.lazy";
 
 // In-flow editor panel — replaces the photos sidebar's filter
@@ -177,59 +175,6 @@ const UnlockRow = styled.label`
   padding: 6px 8px;
   border: 1px dashed var(--inactive-color);
   border-radius: 4px;
-  font-size: 0.85em;
-  color: var(--inactive-color);
-`;
-// Country picker: typeahead-style select using the lang store's
-// localised country dictionary (reused from the public viewer's
-// flag rendering, no new dependency). Stores the alpha-2 code on
-// the form; renders "<flag> Country name (CC)" in the dropdown.
-const CountrySelectRoot = styled.div`
-  position: relative;
-`;
-const CountryDropdown = styled.div`
-  position: absolute;
-  top: 100%;
-  left: 0;
-  right: 0;
-  /* Above leaflet's internal panes (tile 200, overlay 400,
-     marker 600, popup 700) AND its control containers
-     (.leaflet-top / -bottom wrap the +/- zoom buttons at
-     z-index: 1000). Stay below the modal layer (2000) used by
-     the bulk-action backdrops. */
-  z-index: 1100;
-  margin-top: 2px;
-  max-height: 240px;
-  overflow-y: auto;
-  background: var(--primary-background);
-  color: var(--primary-color);
-  border: 1px solid var(--inactive-color);
-  border-radius: 4px;
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.25);
-`;
-const CountryOption = styled.button<{ $active: boolean }>`
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  width: 100%;
-  padding: 6px 10px;
-  background: ${({ $active }) =>
-    $active ? "var(--header-background)" : "transparent"};
-  color: ${({ $active }) =>
-    $active ? "var(--header-color)" : "var(--primary-color)"};
-  border: none;
-  text-align: left;
-  font: inherit;
-  font-size: 0.9em;
-  cursor: pointer;
-  &:hover {
-    background: var(--header-background);
-    color: var(--header-color);
-  }
-`;
-const CountryCode = styled.span`
-  margin-left: auto;
-  font-family: monospace;
   font-size: 0.85em;
   color: var(--inactive-color);
 `;
@@ -689,180 +634,6 @@ const activeMissing = (searchParams: URLSearchParams): Set<MissingField> => {
   return out;
 };
 
-// Inline typeahead for the country field. Reuses the lang store's
-// `countryData` (the same i18n-iso-countries instance the public
-// viewer uses for flag rendering) — no extra network call, no new
-// dependency. Stores the alpha-2 code on the form; the dropdown
-// renders localised names. Empty filter shows all countries
-// alphabetised; typing narrows.
-interface CountrySelectProps {
-  value: string;
-  onChange: (code: string) => void;
-  highlight?: boolean;
-  disabled?: boolean;
-}
-const CountrySelect = ({
-  value,
-  onChange,
-  highlight,
-  disabled,
-}: CountrySelectProps): React.ReactElement => {
-  const { t } = useTranslation();
-  const lang = useLangStore((s) => s.lang);
-  const countryData = useLangStore((s) => s.countryData);
-  const [filter, setFilter] = React.useState("");
-  const [open, setOpen] = React.useState(false);
-  const rootRef = React.useRef<HTMLDivElement>(null);
-
-  React.useEffect(() => {
-    if (!open) return;
-    const onPointer = (e: PointerEvent) => {
-      if (rootRef.current && !rootRef.current.contains(e.target as Node)) {
-        setOpen(false);
-      }
-    };
-    const onKey = (e: KeyboardEvent) => {
-      if (e.key === "Escape") setOpen(false);
-    };
-    document.addEventListener("pointerdown", onPointer);
-    document.addEventListener("keydown", onKey);
-    return () => {
-      document.removeEventListener("pointerdown", onPointer);
-      document.removeEventListener("keydown", onKey);
-    };
-  }, [open]);
-
-  const names = React.useMemo(() => {
-    if (!countryData) return {} as Record<string, string>;
-    return countryData.getNames(lang) as Record<string, string>;
-  }, [countryData, lang]);
-
-  const matches = React.useMemo(() => {
-    const needle = filter.trim().toLowerCase();
-    const entries = Object.entries(names).map(([code, name]) => ({
-      code: code.toLowerCase(),
-      name,
-    }));
-    entries.sort((a, b) => a.name.localeCompare(b.name));
-    if (!needle) return entries;
-    return entries.filter(
-      (e) =>
-        e.name.toLowerCase().includes(needle) ||
-        e.code.toLowerCase().includes(needle)
-    );
-  }, [names, filter]);
-
-  const display = React.useMemo(() => {
-    if (!value) return "";
-    if (isCountrySentinel(value)) {
-      return String(t("country-sentinel-label"));
-    }
-    const upper = value.toUpperCase();
-    const name = names[upper];
-    return name ? `${name} (${upper.toUpperCase()})` : upper.toUpperCase();
-  }, [value, names, t]);
-
-  // Whether the operator's typed filter matches the sentinel
-  // option. Lets typing `no country` / "international" / `xx`
-  // surface the pinned top option even when it doesn't fit the
-  // ISO-code-or-name shape the rest of the matcher checks.
-  const sentinelLabel = String(t("country-sentinel-label"));
-  const sentinelMatches = React.useMemo(() => {
-    const needle = filter.trim().toLowerCase();
-    if (!needle) return true;
-    return (
-      sentinelLabel.toLowerCase().includes(needle) ||
-      "xx".includes(needle)
-    );
-  }, [filter, sentinelLabel]);
-
-  const choose = (code: string) => {
-    onChange(code);
-    setFilter("");
-    setOpen(false);
-  };
-
-  const openDropdown = () => {
-    // Pre-populate the filter with the current country's localised
-    // name so the operator sees the current selection on first
-    // open, narrowed to a tight list. They can edit or clear the
-    // filter to browse; no scroll-into-view needed because the
-    // pre-population keeps the match set small enough to fit.
-    if (value) {
-      if (isCountrySentinel(value)) {
-        setFilter(sentinelLabel);
-      } else {
-        const upper = value.toUpperCase();
-        setFilter(names[upper] ?? upper);
-      }
-    }
-    setOpen(true);
-  };
-
-  return (
-    <CountrySelectRoot ref={rootRef}>
-      <Input
-        type="text"
-        value={open ? filter : display}
-        placeholder={String(t("manage-photo-country-filter-placeholder"))}
-        onFocus={openDropdown}
-        onChange={(e) => {
-          setFilter(e.target.value);
-          if (!open) setOpen(true);
-        }}
-        $highlight={!!highlight}
-        disabled={disabled}
-      />
-      {open && (
-        <CountryDropdown role="listbox">
-          {value && (
-            <CountryOption
-              type="button"
-              $active={false}
-              onClick={() => choose("")}
-              title={String(t("manage-photo-country-clear"))}
-            >
-              <span>—</span>
-              <span>{t("manage-photo-country-clear")}</span>
-            </CountryOption>
-          )}
-          {sentinelMatches && (
-            <CountryOption
-              type="button"
-              $active={isCountrySentinel(value)}
-              onClick={() => choose(COUNTRY_SENTINEL)}
-              title={sentinelLabel}
-            >
-              <span>{sentinelLabel}</span>
-              <CountryCode>XX</CountryCode>
-            </CountryOption>
-          )}
-          {matches.length === 0 && !sentinelMatches && (
-            <CountryOption type="button" $active={false} disabled>
-              <span>{t("manage-photo-country-no-match")}</span>
-            </CountryOption>
-          )}
-          {matches.map((entry) => {
-            const codeUpper = entry.code.toUpperCase();
-            const selected =
-              value && value.toUpperCase() === codeUpper;
-            return (
-              <CountryOption
-                key={entry.code}
-                type="button"
-                $active={!!selected}
-                onClick={() => choose(entry.code)}
-              >
-                <span>{entry.name}</span>
-                <CountryCode>{codeUpper}</CountryCode>
-              </CountryOption>
-            );
-          })}
-        </CountryDropdown>
-      )}
-    </CountrySelectRoot>
-  );
-};
 
 const PhotoDrawer = (): React.ReactElement => {
   const { t } = useTranslation();
